refactor non-db fixtures
