+++ services/api-server/tests/unit/api_solvers/conftest.py
@@ -7,43 +7,10 @@ from typing import Iterator
 
 import pytest
 import respx
-from faker import Faker
 from fastapi import FastAPI
 from pytest_simcore.helpers import catalog_data_fakers
-from requests.auth import HTTPBasicAuth
 from respx import MockRouter
 from simcore_service_api_server.core.settings import ApplicationSettings
-
-
-@pytest.fixture
-def auth(mocker, app: FastAPI, faker: Faker) -> HTTPBasicAuth:
-    # mock engine if db was not init
-    if app.state.settings.API_SERVER_POSTGRES is None:
-
-        engine = mocker.Mock()
-        engine.minsize = 1
-        engine.size = 10
-        engine.freesize = 3
-        engine.maxsize = 10
-        app.state.engine = engine
-
-    # patch authentication entry in repo
-    faker_user_id = faker.pyint()
-
-    # NOTE: here, instead of using the database, we patch repositories interface
-    mocker.patch(
-        "simcore_service_api_server.db.repositories.api_keys.ApiKeysRepository.get_user_id",
-        return_value=faker_user_id,
-    )
-    mocker.patch(
-        "simcore_service_api_server.db.repositories.users.UsersRepository.get_user_id",
-        return_value=faker_user_id,
-    )
-    mocker.patch(
-        "simcore_service_api_server.db.repositories.users.UsersRepository.get_email_from_user_id",
-        return_value=faker.email(),
-    )
-    return HTTPBasicAuth(faker.word(), faker.password())
 
 
 @pytest.fixture
